mid stream commit

+++ src/server/pvfs2-server.h
@@ -610,7 +610,7 @@ struct PINT_server_getattr_op
     PVFS_handle handle;
     PVFS_fs_id fs_id;
     PVFS_ds_attributes dirdata_ds_attr;
-    uint32_t attrmask;
+    uint64_t attrmask;
     PVFS_ds_keyval_handle_info keyval_handle_info;
     int num_dfiles_req;
     PVFS_handle *mirror_dfile_status_array;
